Update Examples structure
