+++ Examples/GreedyKitExample/GreedyKitExample/GreedyKitExampleApp.swift
@@ -1,0 +1,17 @@
+//
+//  GreedyKitExampleApp.swift
+//  GreedyKitExample
+//
+//  Created by Igor Belov on 02.07.2025.
+//
+
+import SwiftUI
+
+@main
+struct GreedyKitExampleApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
