--- Examples/SwiftUIGreedyKit/SwiftUIGreedyKit/SwiftUIGreedyKitApp.swift
@@ -1,17 +1,0 @@
-//
-//  SwiftUIGreedyKitApp.swift
-//  SwiftUIGreedyKit
-//
-//  Created by Igor Belov on 11.09.2022.
-//
-
-import SwiftUI
-
-@main
-struct SwiftUIGreedyKitApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
