+++ Sources/GreedyKit/Core/Utils/ContentGravity.swift
@@ -34,4 +34,3 @@ extension ContentGravity {
         }
     }
 }
-
